test: add test for start_runtime
