+++ vsmtp-rule-engine/src/tests/engine/mod.rs
@@ -61,7 +61,7 @@ fn test_rule_state() {
         .with_default_smtp_codes()
         .without_auth()
         .with_default_app()
-        .with_vsl("./src/empty_main.vsl")
+        .with_vsl("./src/tests/empty_main.vsl")
         .with_default_app_logs()
         .without_services()
         .with_system_dns()
