+++ vsmtp-test/src/config.rs
@@ -22,7 +22,7 @@ pub fn local_test() -> Config {
         .with_default_smtp_codes()
         .without_auth()
         .with_default_app()
-        .with_vsl("src/empty_main.vsl")
+        .with_vsl("src/tests/empty_main.vsl")
         .with_default_app_logs()
         .without_services()
         .with_system_dns()
