+++ vsmtp-test/src/tests/empty_main.vsl
