+++ vsmtp-test/src/tests/tls/mod.rs
@@ -33,7 +33,7 @@ pub fn get_tls_config() -> Config {
         .with_default_smtp_codes()
         .without_auth()
         .with_default_app()
-        .with_vsl("./src/empty_main.vsl")
+        .with_vsl("./src/tests/empty_main.vsl")
         .with_default_app_logs()
         .without_services()
         .with_system_dns()
