+++ vsmtp-test/src/tests/tls/tunneled_with_auth.rs
@@ -24,7 +24,7 @@ fn get_tls_auth_config() -> Config {
         .with_default_smtp_codes()
         .with_safe_auth(true, -1)
         .with_default_app()
-        .with_vsl("./src/empty_main.vsl")
+        .with_vsl("./src/tests/empty_main.vsl")
         .with_default_app_logs()
         .without_services()
         .with_system_dns()
